refactor perspective name
