+++ codbex-products/codbex-products.model
@@ -1692,8 +1692,8 @@
                 "name": "Catalogue",
                 "navigationPath": "/Home",
                 "perspectiveIcon": "/services/web/resources/unicons/book-open.svg",
-                "perspectiveLabel": "Catalogue",
-                "perspectiveName": "Catalogue",
+                "perspectiveLabel": "Catalogues",
+                "perspectiveName": "Catalogues",
                 "perspectiveOrder": "300",
                 "perspectiveRole": "",
                 "title": "Catalogue",
@@ -1724,8 +1724,8 @@
                 "role": {}
             },
             {
-                "name": "Catalogue",
-                "label": "Catalogue",
+                "name": "Catalogues",
+                "label": "Catalogues",
                 "icon": "/services/web/resources/unicons/book-open.svg",
                 "order": "300",
                 "role": {}
