--- codbex-products/gen/dao/Catalogue/Catalogue.extensionpoint
@@ -1,4 +1,0 @@
-{
-    "name": "codbex-products-Catalogue-Catalogue",
-    "description": "Extension Point for the codbex-products-Catalogue-Catalogue entity"
-}
