+++ codbex-products/gen/dao/Catalogues/Catalogue.extensionpoint
@@ -1,0 +1,4 @@
+{
+    "name": "codbex-products-Catalogues-Catalogue",
+    "description": "Extension Point for the codbex-products-Catalogues-Catalogue entity"
+}
