--- codbex-products/gen/ui/Catalogue/Catalogue/dialog-window/view.extension
@@ -1,1 +1,0 @@
-{"module":"codbex-products/gen/ui/Catalogue/Catalogue/dialog-window/view.js","extensionPoint":"codbex-products-dialog-window","description":"codbex-products - Application Dialog Window"}
