--- codbex-products/gen/ui/Catalogue/Catalogue/tile-portal.extension
@@ -1,1 +1,0 @@
-{"module":"codbex-products/gen/ui/Catalogue/Catalogue/tile.js","extensionPoint":"portal-tile","description":"codbex-products - Application Tile"}
