--- codbex-products/gen/ui/Catalogue/Catalogue/tile.extension
@@ -1,1 +1,0 @@
-{"module":"codbex-products/gen/ui/Catalogue/Catalogue/tile.js","extensionPoint":"codbex-products-tile","description":"codbex-products - Application Tile"}
