--- codbex-products/gen/ui/Catalogue/Catalogue/view.extension
@@ -1,1 +1,0 @@
-{"module":"codbex-products/gen/ui/Catalogue/Catalogue/view.js","extensionPoint":"codbex-products-view","description":"codbex-products - Application View"}
