--- codbex-products/gen/ui/Catalogue/perspective-portal.extension
@@ -1,1 +1,0 @@
-{"module":"codbex-products/gen/ui/Catalogue/perspective.js","extensionPoint":"portal","description":"codbex-products - Perspective - Catalogue"}
