--- codbex-products/gen/ui/Catalogue/perspective.extension
@@ -1,1 +1,0 @@
-{"module":"codbex-products/gen/ui/Catalogue/perspective.js","extensionPoint":"codbex-products","description":"codbex-products - Perspective - Catalogue"}
