+++ codbex-products/gen/ui/Catalogues/Catalogue/dialog-window/view.extension
@@ -1,0 +1,1 @@
+{"module":"codbex-products/gen/ui/Catalogues/Catalogue/dialog-window/view.js","extensionPoint":"codbex-products-dialog-window","description":"codbex-products - Application Dialog Window"}
