+++ codbex-products/gen/ui/Catalogues/Catalogue/tile-portal.extension
@@ -1,0 +1,1 @@
+{"module":"codbex-products/gen/ui/Catalogues/Catalogue/tile.js","extensionPoint":"portal-tile","description":"codbex-products - Application Tile"}
