+++ codbex-products/gen/ui/Catalogues/Catalogue/tile.extension
@@ -1,0 +1,1 @@
+{"module":"codbex-products/gen/ui/Catalogues/Catalogue/tile.js","extensionPoint":"codbex-products-tile","description":"codbex-products - Application Tile"}
