+++ codbex-products/gen/ui/Catalogues/Catalogue/view.extension
@@ -1,0 +1,1 @@
+{"module":"codbex-products/gen/ui/Catalogues/Catalogue/view.js","extensionPoint":"codbex-products-view","description":"codbex-products - Application View"}
