+++ codbex-products/gen/ui/Catalogues/Catalogue/view.js
@@ -8,8 +8,8 @@ const viewData = {
     label: "Catalogue",
     factory: "frame",
     region: "center",
-    link: "/services/web/codbex-products/gen/ui/Catalogue/Catalogue/index.html",
-    perspectiveName: "Catalogue",
+    link: "/services/web/codbex-products/gen/ui/Catalogues/Catalogue/index.html",
+    perspectiveName: "Catalogues",
 };
 
 if (typeof exports !== 'undefined') {
