+++ codbex-products/gen/ui/Catalogues/perspective-portal.extension
@@ -1,0 +1,1 @@
+{"module":"codbex-products/gen/ui/Catalogues/perspective.js","extensionPoint":"portal","description":"codbex-products - Perspective - Catalogues"}
