+++ codbex-products/gen/ui/Catalogues/perspective.extension
@@ -1,0 +1,1 @@
+{"module":"codbex-products/gen/ui/Catalogues/perspective.js","extensionPoint":"codbex-products","description":"codbex-products - Perspective - Catalogues"}
